Revert "Convert GridMetricCalculator, GridRange, GridUtils to TypeScript (#302)"
This reverts commit 3b46b7ac3ce677d6d23adcf90aab9f753bda916d.
