+++ packages/grid/src/GridMetricCalculator.test.js
@@ -1,0 +1,38 @@
+import GridMetricCalculator from './GridMetricCalculator';
+
+describe('trimMap', () => {
+  function makeMap(low = 0, high = 10) {
+    const map = new Map();
+    for (let i = low; i <= high; i += 1) {
+      map.set(i, i);
+    }
+    return map;
+  }
+
+  function expectResult(map, expectedMap, cacheSize = 10, targetSize = 5) {
+    GridMetricCalculator.trimMap(map, cacheSize, targetSize);
+    expect(map.size).toEqual(expectedMap.size);
+    const iter = map.entries();
+    const expectedIter = expectedMap.entries();
+    let iterValue = iter.next();
+    let expectedIterValue = expectedIter.next();
+    while (!iterValue.done) {
+      expect(iterValue.value[0]).toEqual(expectedIterValue.value[0]);
+      expect(iterValue.value[1]).toEqual(expectedIterValue.value[1]);
+      iterValue = iter.next();
+      expectedIterValue = expectedIter.next();
+    }
+  }
+
+  it('does not change map if within trim size', () => {
+    expectResult(new Map(), new Map());
+    expectResult(makeMap(0, 0), makeMap(0, 0));
+    expectResult(makeMap(0, 9), makeMap(0, 9));
+  });
+
+  it('trims map if larger than cache size', () => {
+    expectResult(makeMap(0, 10), makeMap(6, 10));
+    expectResult(makeMap(0, 100), makeMap(96, 100));
+    expectResult(makeMap(0, 100), makeMap(51, 100), 100, 50);
+  });
+});
