+++ packages/grid/src/GridRange.js
@@ -1,51 +1,12 @@
-export type GridRangeIndex = number | null;
-type LeftIndex = GridRangeIndex;
-type RightIndex = GridRangeIndex;
-type TopIndex = GridRangeIndex;
-type BottomIndex = GridRangeIndex;
-
-export type GridCell = { column: number; row: number };
-
-export interface BoundedGridRange extends GridRange {
-  startColumn: number;
-  startRow: number;
-  endColumn: number;
-  endRow: number;
-}
-
-// Also exported via GridRange.SELECTION_DIRECTION
-export enum SELECTION_DIRECTION {
-  DOWN = 'DOWN',
-  UP = 'UP',
-  LEFT = 'LEFT',
-  RIGHT = 'RIGHT',
-}
-
 class GridRange {
-  startColumn: GridRangeIndex;
+  static SELECTION_DIRECTION = Object.freeze({
+    DOWN: 'DOWN',
+    UP: 'UP',
+    LEFT: 'LEFT',
+    RIGHT: 'RIGHT',
+  });
 
-  startRow: GridRangeIndex;
-
-  endColumn: GridRangeIndex;
-
-  endRow: GridRangeIndex;
-
-  static SELECTION_DIRECTION = SELECTION_DIRECTION;
-
-  /**
-   * Returns a normalized array of indexes ensuring left <= right and top <= bottom
-   * @param startColumn Start column index
-   * @param startRow Start row index
-   * @param endColumn End column index
-   * @param endRow End row index
-   * @returns Array containing normalized indexes [left, top, right, bottom]
-   */
-  static normalize(
-    startColumn: GridRangeIndex,
-    startRow: GridRangeIndex,
-    endColumn: GridRangeIndex,
-    endRow: GridRangeIndex
-  ): [LeftIndex, TopIndex, RightIndex, BottomIndex] {
+  static normalize(startColumn, startRow, endColumn, endRow) {
     let left = startColumn;
     let top = startRow;
     let right = endColumn;
@@ -64,93 +25,44 @@ class GridRange {
     return [left, top, right, bottom];
   }
 
-  /**
-   * Makes a GridRange ensuring startColumn <= endColumn, startRow <= endRow
-   * @param startColumn Start column index
-   * @param startRow Start row index
-   * @param endColumn End column index
-   * @param endRow End row index
-   * @returns Normalized GridRange
-   */
-  static makeNormalized(
-    startColumn: GridRangeIndex,
-    startRow: GridRangeIndex,
-    endColumn: GridRangeIndex,
-    endRow: GridRangeIndex
-  ): GridRange {
-    return new GridRange(
-      ...GridRange.normalize(startColumn, startRow, endColumn, endRow)
-    );
+  /** Make a GridRange, but ensure startColumn <= endColumn, startRow <= endRow */
+  static makeNormalized(...coordinates) {
+    return new GridRange(...GridRange.normalize(...coordinates));
   }
 
-  /**
-   * Creates a GridRange representing a single cell
-   * @param column Column index
-   * @param row Row index
-   * @returns GridRange representing the cell
-   */
-  static makeCell(column: GridRangeIndex, row: GridRangeIndex): GridRange {
+  static makeCell(column, row) {
     return new GridRange(column, row, column, row);
   }
 
-  /**
-   * Creates a GridRange representing an infinite length column
-   * @param column Column index
-   * @returns GridRange representing the column
-   */
-  static makeColumn(column: GridRangeIndex): GridRange {
+  static makeColumn(column) {
     return new GridRange(column, null, column, null);
   }
 
-  /**
-   * Creates a GridRange representing an infinite length row
-   * @param row Row index
-   * @returns GridRange representing the row
-   */
-  static makeRow(row: GridRangeIndex): GridRange {
+  static makeRow(row) {
     return new GridRange(null, row, null, row);
   }
 
-  /**
-   * Returns the minimum value between 2 range indexes or null if at least 1 is null
-   * @param index1 First grid range index
-   * @param index2 Second grid range index
-   * @returns Minimum index or null if either index is null
-   */
-  static minOrNull(
-    index1: GridRangeIndex,
-    index2: GridRangeIndex
-  ): number | null {
-    if (index1 == null || index2 == null) {
+  static minOrNull(value1, value2) {
+    if (value1 == null || value2 == null) {
       return null;
     }
 
-    return Math.min(index1, index2);
+    return Math.min(value1, value2);
   }
 
-  /**
-   * Returns the maximum value between 2 range indexes or null if at least 1 is null
-   * @param index1 First grid range index
-   * @param index2 Second grid range index
-   * @returns Maximum index or null if either index is null
-   */
-  static maxOrNull(
-    index1: GridRangeIndex,
-    index2: GridRangeIndex
-  ): number | null {
-    if (index1 == null || index2 == null) {
+  static maxOrNull(value1, value2) {
+    if (value1 == null || value2 == null) {
       return null;
     }
 
-    return Math.max(index1, index2);
+    return Math.max(value1, value2);
   }
 
   /**
    * Consolidate the passed in ranges to the minimum set, merging overlapping ranges.
-   * @param ranges The ranges to consolidate
-   * @returns Consolidated ranges
+   * @param {[GridRange]} ranges The ranges to consolidate
    */
-  static consolidate(ranges: GridRange[]): GridRange[] {
+  static consolidate(ranges) {
     const result = ranges.slice();
 
     let wasModified = true;
@@ -226,69 +138,47 @@ class GridRange {
     return result;
   }
 
-  /**
-   * Checks if the 1-D ranges between 2 index pairs overlap or are continuous.
-   * For example ranges [0, 1] and [2, 3] are continuous and will return true.
-   * [0, 1] and [1, 3] overlap and return true.
-   * [0, 1] and [3, 4] do not overlap and have a gap so this will return false.
-   * @param start1 Start of 1st range
-   * @param end1 End of 1st range
-   * @param start2 Start of 2nd range
-   * @param end2 End of 2nd range
-   * @returns True if the ranges overlap or touch, else false
-   */
-  static isAxisRangeTouching(
-    start1: GridRangeIndex,
-    end1: GridRangeIndex,
-    start2: GridRangeIndex,
-    end2: GridRangeIndex
-  ): boolean {
-    if (start1 == null) {
-      if (end1 == null) {
+  static isAxisRangeTouching(start, end, otherStart, otherEnd) {
+    if (start == null) {
+      if (end == null) {
         return true;
       }
 
-      if (start2 == null) {
+      if (otherStart == null) {
         return true;
       }
 
-      return start2 <= end1 + 1;
+      return otherStart <= end + 1;
     }
 
-    if (end1 == null) {
-      if (end2 == null) {
+    if (end == null) {
+      if (otherEnd == null) {
         return true;
       }
 
-      return end2 >= start1 - 1;
+      return otherEnd >= start - 1;
     }
 
-    if (start2 == null) {
-      if (end2 == null) {
+    if (otherStart == null) {
+      if (otherEnd == null) {
         return true;
       }
 
-      return start1 <= end2 + 1;
+      return start <= otherEnd + 1;
     }
 
-    if (end2 == null) {
-      return end1 >= start2 - 1;
+    if (otherEnd == null) {
+      return end >= otherStart - 1;
     }
 
-    if (start2 >= start1 - 1) {
-      return start2 <= end1 + 1;
+    if (otherStart >= start - 1) {
+      return otherStart <= end + 1;
     }
 
-    return end2 >= start1 - 1;
+    return otherEnd >= start - 1;
   }
 
-  /**
-   * Checks if 2 arrays of ranges are the same ranges
-   * @param ranges1 First array of ranges
-   * @param ranges2 Second array of ranges
-   * @returns True if the arrays contain the same ranges in the same order
-   */
-  static rangeArraysEqual(ranges1: GridRange[], ranges2: GridRange[]): boolean {
+  static rangeArraysEqual(ranges1, ranges2) {
     if (ranges1 === ranges2) {
       return true;
     }
@@ -312,14 +202,11 @@ class GridRange {
 
   /**
    * Get the intersection (overlapping area) of two ranges
-   * @param range One range to check for the intersection
-   * @param otherRange The other range to check for the intersection
-   * @returns Intersection of the two ranges. If they do not intersect, returns `null`.
+   * @param {GridRange} range One range to check for the intersection
+   * @param {GridRange} otherRange The other range to check for the intersection
+   * @returns {GridRange|null} Intersection of the two ranges. If they do not intersect, returns `null`.
    */
-  static intersection(
-    range: GridRange,
-    otherRange: GridRange
-  ): GridRange | null {
+  static intersection(range, otherRange) {
     if (range.equals(otherRange)) {
       return range;
     }
@@ -343,8 +230,8 @@ class GridRange {
         : endRow ?? otherRange.endRow;
 
     if (
-      (startColumn != null && startColumn > (endColumn ?? -1)) ||
-      (startRow != null && startRow > (endRow ?? -1))
+      (startColumn != null && startColumn > endColumn) ||
+      (startRow != null && startRow > endRow)
     ) {
       return null;
     }
@@ -353,15 +240,11 @@ class GridRange {
   }
 
   /**
-   * Subtracts 1 range from another
-   * @param range The range to be subtracted from
-   * @param subtractRange The range to subtract from within this range
-   * @returns The ranges needed to represent the remaining
+   * @param {GridRange} range The range to be subtracted from
+   * @param {GridRange} subtractRange The range to subtract from within this range
+   * @returns {GridRange[]} The ranges needed to represent the remaining
    */
-  static subtractFromRange(
-    range: GridRange,
-    subtractRange: GridRange
-  ): GridRange[] {
+  static subtractFromRange(range, subtractRange) {
     const result = [];
 
     // Make it a little easier by finding only the part the subtraction range intersects
@@ -436,15 +319,12 @@ class GridRange {
 
   /**
    * Subtract a range from multiple ranges
-   * @param ranges The ranges to be subtracted from
-   * @param subtractRange The range to subtract from within these ranges
-   * @returns The ranges needed to represent the remaining
+   * @param {GridRange[]} ranges The ranges to be subtracted from
+   * @param {GridRange} subtractRange The range to subtract from within these ranges
+   * @returns {GridRange[]} The ranges needed to represent the remaining
    */
-  static subtractFromRanges(
-    ranges: GridRange[],
-    subtractRange: GridRange
-  ): GridRange[] {
-    const result: GridRange[] = [];
+  static subtractFromRanges(ranges, subtractRange) {
+    const result = [];
     for (let i = 0; i < ranges.length; i += 1) {
       result.push(...GridRange.subtractFromRange(ranges[i], subtractRange));
     }
@@ -454,14 +334,11 @@ class GridRange {
 
   /**
    * Subtract multiple ranges from multiple ranges
-   * @param ranges The ranges to be subtracted from
-   * @param subtractRanges The ranges to subtract from within these ranges
-   * @returns The ranges needed to represent the remaining
+   * @param {GridRange[]} ranges The ranges to be subtracted from
+   * @param {GridRange[]} subtractRanges The ranges to subtract from within these ranges
+   * @returns {GridRange[]} The ranges needed to represent the remaining
    */
-  static subtractRangesFromRanges(
-    ranges: GridRange[],
-    subtractRanges: GridRange[]
-  ): GridRange[] {
+  static subtractRangesFromRanges(ranges, subtractRanges) {
     if (!subtractRanges || subtractRanges.length === 0) {
       return ranges;
     }
@@ -476,10 +353,10 @@ class GridRange {
 
   /**
    * Test if a given range is bounded (all values are non-null)
-   * @param range The range to test
-   * @returns True if this range is bounded, false otherwise
+   * @param {GridRange} range The range to test
+   * @returns {boolean} True if this range is bounded, false otherwise
    */
-  static isBounded(range: GridRange): range is BoundedGridRange {
+  static isBounded(range) {
     return (
       range.startRow != null &&
       range.startColumn != null &&
@@ -492,16 +369,12 @@ class GridRange {
    * Converts any GridRange passed in that is a full row or column selection to be bound
    * to the `columnCount` and `rowCount` passed in
    *
-   * @param range The range to get the bounded range of
-   * @param columnCount The number of columns
-   * @param rowCount The number of rows
-   * @returns The passed in GridRange with any null values filled in
+   * @param {GridRange} range The range to get the bounded range of
+   * @param {number} columnCount The number of columns
+   * @param {number} rowCount The number of rows
+   * @returns {GridRange} The passed in GridRange with any null values filled in
    */
-  static boundedRange(
-    range: GridRange,
-    columnCount: number,
-    rowCount: number
-  ): GridRange {
+  static boundedRange(range, columnCount, rowCount) {
     if (GridRange.isBounded(range)) {
       return range;
     }
@@ -517,32 +390,24 @@ class GridRange {
   /**
    * Converts the GridRanges passed in to be bound to the `columnCount` and `rowCount` passed in
    *
-   * @param ranges The ranges to get the bounded ranges of
-   * @param columnCount The number of columns
-   * @param rowCount The number of rows
-   * @returns The passed in GridRange with any null values filled in
+   * @param {GridRange[]} ranges The ranges to get the bounded ranges of
+   * @param {number} columnCount The number of columns
+   * @param {number} rowCount The number of rows
+   * @returns {GridRange} The passed in GridRange with any null values filled in
    */
-  static boundedRanges(
-    ranges: GridRange[],
-    columnCount: number,
-    rowCount: number
-  ): GridRange[] {
+  static boundedRanges(ranges, columnCount, rowCount) {
     return ranges.map(r => GridRange.boundedRange(r, columnCount, rowCount));
   }
 
   /**
    * Offsets a GridRange by the specified amount in the x and y directions
    *
-   * @param range The range to offset
-   * @param columnOffset The number of columns to offset
-   * @param rowOffset The number of rows to offset
-   * @returns The new grid range offset from the original
+   * @param {GridRange} range The range to offset
+   * @param {number} columnOffset The number of columns to offset
+   * @param {number} rowOffset The number of rows to offset
+   * @returns {GridRange} The new grid range offset from the original
    */
-  static offset(
-    range: GridRange,
-    columnOffset: number,
-    rowOffset: number
-  ): GridRange {
+  static offset(range, columnOffset, rowOffset) {
     return new GridRange(
       range.startColumn != null ? range.startColumn + columnOffset : null,
       range.startRow != null ? range.startRow + rowOffset : null,
@@ -553,18 +418,18 @@ class GridRange {
 
   /**
    * Get the next cell given the selected ranges and the current cell
-   * @param ranges The selected bounded ranges within the grid
-   * @param column The cursor column, or null if none focused
-   * @param row The cursor row, or null if none focused
-   * @param direction The direction in which to select next
-   * @returns The next cell to focus, or null if there should be no more focus
+   * @param {GridRange[]} ranges The selected bounded ranges within the grid
+   * @param {number|null} column The cursor column, or null if none focused
+   * @param {number|null} row The cursor row, or null if none focused
+   * @param {SELECTION_DIRECTION} direction The direction in which to select next
+   * @returns {Cell} The next cell to focus, or null if there should be no more focus
    */
   static nextCell(
-    ranges: GridRange[],
-    column: GridRangeIndex = null,
-    row: GridRangeIndex = null,
+    ranges,
+    column = null,
+    row = null,
     direction = GridRange.SELECTION_DIRECTION.DOWN
-  ): GridCell | null {
+  ) {
     if (ranges.length === 0) {
       return null;
     }
@@ -605,10 +470,10 @@ class GridRange {
 
   /**
    * Count the number of cells in the provided grid ranges
-   * @param ranges The ranges to count the rows of
-   * @returns The number of cells in the ranges, or `NaN` if any of the ranges were unbounded
+   * @param {GridRange[]} ranges The ranges to count the rows of
+   * @returns {number|NaN} The number of cells in the ranges, or `NaN` if any of the ranges were unbounded
    */
-  static cellCount(ranges: GridRange[]): number {
+  static cellCount(ranges) {
     return ranges.reduce(
       (cellCount, range) =>
         cellCount +
@@ -620,10 +485,10 @@ class GridRange {
 
   /**
    * Count the number of rows in the provided grid ranges
-   * @param ranges The ranges to count the rows of
-   * @returns The number of rows in the ranges, or `NaN` if any of the ranges were unbounded
+   * @param {GridRange[]} ranges The ranges to count the rows of
+   * @returns {number|NaN} The number of rows in the ranges, or `NaN` if any of the ranges were unbounded
    */
-  static rowCount(ranges: GridRange[]): number {
+  static rowCount(ranges) {
     return ranges.reduce(
       (rowCount, range) =>
         rowCount + (range.endRow ?? NaN) - (range.startRow ?? NaN) + 1,
@@ -633,10 +498,10 @@ class GridRange {
 
   /**
    * Count the number of columns in the provided grid ranges
-   * @param ranges The ranges to count the columns of
-   * @returns The number of columns in the ranges, or `NaN` if any of the ranges were unbounded
+   * @param {GridRange[]} ranges The ranges to count the columns of
+   * @returns {number|NaN} The number of columns in the ranges, or `NaN` if any of the ranges were unbounded
    */
-  static columnCount(ranges: GridRange[]): number {
+  static columnCount(ranges) {
     return ranges.reduce(
       (columnCount, range) =>
         columnCount + (range.endColumn ?? NaN) - (range.startColumn ?? NaN) + 1,
@@ -646,16 +511,12 @@ class GridRange {
 
   /**
    * Check if the provided ranges contain the provided cell
-   * @param ranges The ranges to check
-   * @param column The column index
-   * @param row The row index
-   * @returns True if the cell is within the provided ranges, false otherwise.
+   * @param {GridRange[]} ranges The ranges to check
+   * @param {number} column The column index
+   * @param {number} row The row index
+   * @returns {boolean} True if the cell is within the provided ranges, false otherwise.
    */
-  static containsCell(
-    ranges: GridRange[],
-    column: number,
-    row: number
-  ): boolean {
+  static containsCell(ranges, column, row) {
     for (let i = 0; i < ranges.length; i += 1) {
       const range = ranges[i];
       if (range.containsCell(column, row)) {
@@ -667,38 +528,28 @@ class GridRange {
 
   /**
    * Iterate through each cell in the provided ranges
-   * @param ranges The ranges to iterate through
+   * @param {GridRange[]} ranges The ranges to iterate through
    * @param {(column: number, row: number, index: number) => void} callback The callback to execute. `index` is the index within that range
    * @param {GridRange.SELECTION_DIRECTION} direction The direction to iterate in
    */
   static forEachCell(
-    ranges: GridRange[],
-    callback: (column: number, row: number, index: number) => void,
+    ranges,
+    callback,
     direction = GridRange.SELECTION_DIRECTION.RIGHT
-  ): void {
+  ) {
     for (let i = 0; i < ranges.length; i += 1) {
       ranges[i].forEach(callback, direction);
     }
   }
 
-  constructor(
-    startColumn: GridRangeIndex,
-    startRow: GridRangeIndex,
-    endColumn: GridRangeIndex,
-    endRow: GridRangeIndex
-  ) {
+  constructor(startColumn, startRow, endColumn, endRow) {
     this.startColumn = startColumn;
     this.startRow = startRow;
     this.endColumn = endColumn;
     this.endRow = endRow;
   }
 
-  /**
-   * Checks if the provided range is equivalent to this range (same start and end column/row indexes)
-   * @param other Grid range to check against
-   * @returns True if the ranges cover the same area
-   */
-  equals(other: GridRange): boolean {
+  equals(other) {
     return (
       this.startColumn === other.startColumn &&
       this.startRow === other.startRow &&
@@ -707,12 +558,8 @@ class GridRange {
     );
   }
 
-  /**
-   * Checks if this GridRange contains another range
-   * @param other The range to check
-   * @returns True if this GridRange completely contains `other`
-   * */
-  contains(other: GridRange): boolean {
+  /** @returns {boolean} true if this GridRange completely contains `other` */
+  contains(other) {
     return (
       (this.startColumn == null ||
         (other.startColumn != null && this.startColumn <= other.startColumn)) &&
@@ -727,11 +574,11 @@ class GridRange {
 
   /**
    * Check if the provided cell is in this range
-   * @param column The column to check
-   * @param row The row to check
-   * @returns True if this cell is within this range
+   * @param {number} column The column to check
+   * @param {number} row The row to check
+   * @returns {boolean} True if this cell is within this range
    */
-  containsCell(column: GridRangeIndex, row: GridRangeIndex): boolean {
+  containsCell(column, row) {
     if (column == null || row == null) {
       return false;
     }
@@ -744,13 +591,8 @@ class GridRange {
     );
   }
 
-  /**
-   * Check if the provided range touches (or overlaps) this GridRange
-   * Effectively checks if the 2 ranges could be represented by 1 continuous range
-   * @param other The range to check
-   * @returns True if this GridRange touches `other`
-   * */
-  touches(other: GridRange): boolean {
+  /** @returns {boolean} true if this GridRange touches `other` */
+  touches(other) {
     return (
       GridRange.isAxisRangeTouching(
         this.startRow,
@@ -768,21 +610,20 @@ class GridRange {
   }
 
   /**
-   * Subtracts a range from this range
-   * @param other The range to deselect from within this range
-   * @returns The ranges needed to represent the remaining
+   * @param {GridRange} other The range to deselect from within this range
+   * @returns {[GridRange]} The ranges needed to represent the remaining
    */
-  subtract(other: GridRange): GridRange[] {
+  subtract(other) {
     return GridRange.subtractFromRange(this, other);
   }
 
   /**
    * Get the first cell in this range. Throws if this range is unbounded.
    *
-   * @param direction The direction to get the starting cell in. Defaults to DOWN
-   * @returns The first cell in this range in the direction specified
+   * @param {GridRange.SELECTION_DIRECTION?} direction The direction to get the starting cell in. Defaults to DOWN
+   * @returns {{column: number, row: number}} The first cell in this range in the direction specified
    */
-  startCell(direction = GridRange.SELECTION_DIRECTION.DOWN): GridCell {
+  startCell(direction = GridRange.SELECTION_DIRECTION.DOWN) {
     if (!GridRange.isBounded(this)) {
       throw new Error('Cannot get the startCell of an unbounded range');
     }
@@ -811,11 +652,7 @@ class GridRange {
    * @param {SELECTION_DIRECTION} direction The direction to go in
    * @returns {GridCell|null} The next cell in the direction specified, or `null` if at the end of the range
    */
-  nextCell(
-    column: GridRangeIndex,
-    row: GridRangeIndex,
-    direction: SELECTION_DIRECTION
-  ): GridCell | null {
+  nextCell(column, row, direction) {
     if (!GridRange.isBounded(this)) {
       throw new Error('Bounded range required');
     }
@@ -871,18 +708,12 @@ class GridRange {
 
   /**
    * Iterate through each cell in the range
-   * @param callback Callback to execute. `index` is the index within this range
-   * @param direction The direction to iterate in
+   * @param {(column:number, row:number, index:number) => void} callback Callback to execute. `index` is the index within this range
+   * @param {GridRange.SELECTION_DIRECTION} direction The direction to iterate in
    */
-  forEach(
-    callback: (column: number, row: number, index: number) => void,
-    direction = GridRange.SELECTION_DIRECTION.RIGHT
-  ): void {
+  forEach(callback, direction = GridRange.SELECTION_DIRECTION.RIGHT) {
     let i = 0;
-    let {
-      column: c,
-      row: r,
-    }: { column?: number; row?: number } = this.startCell(direction);
+    let { column: c, row: r } = this.startCell(direction);
     while (c != null && r != null) {
       callback(c, r, i);
       i += 1;
