+++ packages/grid/src/mouse-handlers/GridSelectionMouseHandler.js
@@ -86,15 +86,15 @@ class GridSelectionMouseHandler extends GridMouseHandler {
     let deltaX = 0;
     let deltaY = 0;
     if (left < lastLeft) {
-      if (x < dragBounds.x1) {
-        deltaX = x - dragBounds.x1;
+      if (x < dragBounds.x) {
+        deltaX = x - dragBounds.x;
       } else if (x > dragBounds.x2) {
         deltaX = x - dragBounds.x2;
       }
     }
     if (top < lastTop) {
-      if (y < dragBounds.y1) {
-        deltaY = y - dragBounds.y1;
+      if (y < dragBounds.y) {
+        deltaY = y - dragBounds.y;
       } else if (y > dragBounds.y2) {
         deltaY = y - dragBounds.y2;
       }
